Full proof of concept
webNavigation is filtered on the pattern strings entered into the popup
window. These are regexes as defined in the mozilla docs. Triggering a
match closes the offending tab and opens the extension page. The
offending URL is included in the query string.

The extension page includes an override button that provides a one-time
pass to access the forbidden fruit. This is accomplished by writing the
exemption info to disk before attempting to access the URL again.

When the filter callback sees the exemption info in storage.local, it
deletes the exemption information but otherwise does not act, allowing
the filtered URL to load.