+improvements
When an offending URL is detected the extension now redirects to the
extension page rather opening in a new tab. The entry in the tab's
navigation history is then overwritten with the extension page so that
the back button behavior doesn't just loop.

The extension page has been made slightly prettier. Now includes the
offending URL.

The background script now sets an onStart callback, so hopefully the
background scripts will now load the filters on browser start rather
than requiring the browser_action UI to 'save changes'. Remains to be
seen. Will require Mozilla sign off to make the extension persistent.